Add context awareness RBAC permissions
The `policy-server` requires to list well known resources, so
context-aware policies can take decisions based on this information.

While this permissions, and what resources can be listed and watched
will be customizable in the future, right now is a fixed set of
resources:

- Namespaces
- Services
- Ingresses

This list is customizable by overriding the helm chart `values.yaml`.