Revert "feat: Make kubewarden-defaults depend on kubewarden-controller"
This reverts commit 0e2be843ddb90a5725e8a685a2554d7999b07149.

Using this `requires-gvr` doesn't seem to work. Pending resolution.

Instead of properly depending on `kubewarden-controler`, the alternative
could be to set `kubewarden-defaults` as hidden chart in UI, and only
show to the user the `kubewarden-controller` chart, as "Kubewarden".
Then, in Kubewarden UI, ask the user if they want the defaults, and let
the Kubewarden UI install `kubewarden-defaults` chart.