fix: ensure charts can be installed on GKE
GKE clusters have a version number that embeds some GKE metadata. This
metadata is added using the `-gke.<number>`.

Helm gets "confused" by this version and considers that a release
candidate (due to semver rules).

Because of that, prior to this commit, attempting to deploy the
Kubewarden stack on GKE resulted in this error:

```console
Error: INSTALLATION FAILED: chart requires kubeVersion: >= 1.19.0 which is incompatible with Kubernetes v1.23.8-gke.1900
```

This is now fixed

Signed-off-by: Flavio Castelli <fcastelli@suse.com>