review: Correct version of charts
The automation seems to be wrong when bumping the chart version field
for RCs.

Use correct minor version.

Bump kubewarden-crds to `1.3.0-rc6` even if they don't have any changes;
makes it easier to not forget later on that kubewarden-crds version
needs to be changed and charts consume it with
`catalog.cattle.io/auto-install`.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>