Recreate controller when telemetry is enabled.
When telemetry is enabled the OpenTelemetry collector and the telemetry
configuration in the controller deployment is set. However, due a timing
issue the OTEL operator failed to add the sidecar container in the controller
pod because it cannot find the collector. Therefore, it's necessary to
recreate the controller pod after the chart installation.

This commit workaround this by adding a post-install Helm chart hook
which will delete the controller pod. Therefore, the OTEL operator will
be able to find the collector.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>