fix: Ensure installation of PolicyReport CRDs if value is missing or set to "null".
In the Rancher manager environment, an issue arises during updates of
the kubewarden-crds chart where the "installPolicyReportCRDs"
configuration is sometimes set to "null." This results in problems as
the CRDs, which should be installed by default, are not deployed. To
address this issue, we've enhanced the conditional logic to install the
CRDs when the flag is explicitly set to true or when it is missing
altogether.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>