feat: Add .Values.policyServer.telemetry to kubewarden-defaults
Setting this to true will add the necesary `KUBEWARDEN_LOG_FMT` env var
to the default PolicyServer, so it outputs telemetry in OTLP format.

This is needed because the kubewarden controller doesn't take care of
reconciling this env var anymore, allowing users to fine tune it.

For this commit's case, the default PolicyServer, we are fine keeping
the behaviour as it is.

Note that this is not backwards compatible, as
`.Values.policyServer.telemetry` defaults to `false`.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>