feat: Add `make check-common-values` target
This target diffs the contents of `./common-values.yaml` and
`charts/*/chart-values.yaml` for the `global` key.

This is analogous to what `make check-generated-values` does,
but without calling `make generated-values`, nor having an interim
`chart-values.yaml` per chart.

This new target is intended to substitute `make check-generated-values`
and to drop the `chart-values.yaml`. This will simplify contributions
and automation.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>