fix(ci): fix broken updatecli yaml files.
There are two updatecli yaml files missing the `targets` field. Thus,
the execution failed because all the update actions that should be under
`targets` are children of the `conditions` field. This commit that adding
the `targets` field.

This commit also fixes some warning messages from updatecli about the
keys used to find the fields in the Helm chart yaml files.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>