fix: invalid audit scanner container command.
The container command is an array of strings. The `--disable-store` flag
value passed is `false` which fails. Because it is unmarshaled as
boolean. To fix that, the helper code to build the command now only add
the flag if the user want to disable the store. Otherwise, it does not
add the CLI flag that. As the default value for the flag is "false" this
does not change the audit scanner behaviour.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>