ci: Add inputs for manually run "Update helm charts" job
Add UI inputs when workflow triggered as `workflow_dispatch`.

Given that the values can come from `repository_dispatch` (other repos)
or `workflow_dispatch` (UI button), add a new `setvariables` job, that
sets its job outputs to the corresponding values obtained from the
payload of the repository_dispatch or from the inputs of the
workflow_dispatch.
It is needed to use job outputs as is the only way to pass values
between jobs, as jobs are run in their own VM instances and don't share
the env.

Make the following jobs depend on it with `needs: setvariables`.

The javascript script from actions/github-script cannot read job
outputs. Hence, add a step that reads the `needs.setvariables.outputs`
and creates env vars. Consume those env vars in js as
`process.env.foo`.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>