deps: Migrate to rancher/kuberlr-kubectl
Instead of using a hardcoded kubectl version as we did with
kubewarden/kubectl image, migrate to `kuberlr-kubectl`.

kuberlr-kubectl image bundles several kubectl binaries that correspond
with the Kubernetes versions supported by Rancher. This simplifies
airgap installations.

In addition, kuberlr-kubectl will download a matching kubectl binary
if the Kubernetes API server is out of the Kubernetes version skew
policy of the kubectl binaries it bundles. This ensures we always
have a correct kubectl binary.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>