feat(update-hauler): new updatecli policy to update Hauler manifest.
Add a new policy which extract the OCI artifacts and Helm chart version
from the Helm charts and update the Hauler manifest with them.

With the new updatecli policy used exclusively used to update the Hauler
manifest based on the values defined in the Helm chart, the update-deps
policy does not need to update the Hauler manifest. The code for that
has been moved to the new policy. update-deps can focus on the Helm
chart update only.

In a future change both policies will be used together to update every
that is required.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>