feat(audit-scanner): allow storing results with OpenReports.
Updates the audit scanner job allow users to define which CRDs they want
to use to store the audit scan results. The available options are
"openrepots" and "policyreport".

"openreports" option configure the audit scanner to store the results in
the OpenReports CRDs. And, "policyreport" configures the scanner to
store the results in the, now deprecated, Kubernetes policy working
group policy reports CRDs.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>